LoopStudioBot/requirements.txt
# Telegram Bot
python-telegram-bot==21.7

# System monitoring
speedtest-cli==2.1.3
psutil==6.1.0

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.1
requests==2.32.3
